requirements.txt
streamlit
pandas
gspread
google-auth
ifcopenshell
reportlab
qrcode
pillow